backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
exceptiongroup==1.3.0
fastapi==0.115.12
filelock==3.18.0
fsspec==2025.3.2
h11==0.16.0
idna==3.10
Jinja2==3.1.6
llvmlite==0.44.0
MarkupSafe==3.0.2
more-itertools==10.7.0
mpmath==1.3.0
networkx==3.4.2
numba==0.61.2
numpy==1.26.4
openai-whisper==20240930
pydantic==2.11.4
pydantic_core==2.33.2
python-multipart==0.0.20
regex==2024.11.6
requests==2.32.3
sniffio==1.3.1
starlette==0.46.2
sympy==1.14.0
tiktoken==0.9.0
torch==2.2.2
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
